Feature/checkout finished UI enhancement

+++ android/app/src/main/java/org/cryptimeleon/incentive/app/data/BasketRepository.kt
@@ -150,7 +150,7 @@ class BasketRepository(
 
     override suspend fun discardCurrentBasket(delete: Boolean) {
         val basket = basket.first()
-        if (basket != null) {
+        if (delete && basket != null) {
             basketApiService.deleteBasket(basket.basketId)
         }
         basketDao.deleteAllBasketItems()
